+++ packages/react-app/src/components/Header.jsx
@@ -1,14 +1,12 @@
 import { PageHeader } from "antd";
 import React from "react";
 
-// displays a page header
-
 export default function Header() {
   return (
-    <a href="https://github.com/austintgriffith/scaffold-eth" target="_blank" rel="noopener noreferrer">
+    <a href="https://moonshotcollective.space" target="_blank" rel="noopener noreferrer">
       <PageHeader
-        title="🏗 scaffold-eth"
-        subTitle="forkable Ethereum dev stack focused on fast product iteration"
+        title="💰Public Goods Party💰"
+        subTitle="A forkable grant funding toolkit by the moonshot collective!"
         style={{ cursor: "pointer" }}
       />
     </a>
