+++ packages/react-app/src/components/index.js
@@ -21,5 +21,3 @@ export { default as L2Bridge } from "./L2Bridge";
 export { default as NetworkDisplay } from "./NetworkDisplay";
 export { default as FaucetHint } from "./FaucetHint";
 export { default as NetworkSwitch } from "./NetworkSwitch";
-export {default as GrantCreator} from './dGrants/GrantCreator';
-export {default as GrantRoundCreator} from './dGrants/GrantRoundCreator';
