+++ packages/react-app/src/views/index.js
@@ -1,8 +1,6 @@
 export { default as Home } from "./Home";
-export { default as ExampleUI } from "./ExampleUI";
-export { default as Hints } from "./Hints";
 export { default as Subgraph } from "./Subgraph";
-export { default as GrantUI} from "./ExampledGrantUI";
+export { default as GrantCreator} from "./GrantCreator";
 export { default as DetailedGrantView} from "./DetailedGrantView";
 export { default as NameViewer} from "./NameViewer";
 export { default as DonationView} from "./DonationView";
